+++ src/Pre/theme.js
@@ -12,7 +12,8 @@ export const makeTheme = (userTheme = {}) => {
   // the Bootstrap defaults without modifying key, versioned files.
   const v = {};
   const u = userTheme;
-
+  // all var must be grouped in a named theme
+  v['_name'] = u['_name'] || 'bootstrap-styled';
   v['$gray-dark'] = u['$gray-dark'] || '#292b2c';
   // Code