+++ src/Table/theme.js
@@ -13,7 +13,8 @@ export const makeTheme = (userTheme = {}) => {
   // the Bootstrap defaults without modifying key, versioned files.
   const v = {};
   const u = userTheme;
-
+  // all var must be grouped in a named theme
+  v['_name'] = u['_name'] || 'bootstrap-styled';
   // Colors
   //
   // Grayscale and colors for use across Bootstrap.