+++ src/theme/tests/makeExtend.test.js
@@ -4,7 +4,7 @@ describe('makeExtend', () => {
   it('should have makeOriginal defined', () => {
     expect(makeExtend).toBeDefined();
   });
-  it('should have keys starting with $ only', () => {
-    Object.keys(makeExtend()).forEach((key) => expect(key[0]).toEqual('$'));
+  it('should have keys starting with $ or _ only', () => {
+    Object.keys(makeExtend()).forEach((key) => expect(['$', '_'].indexOf(key[0])).toBeGreaterThanOrEqual(0));
   });
 });